// src/types/cssProps.ts
export type ReviewCardCSSProps = {
    reviewCardClassName?: string;
    reviewCardStyle?: React.CSSProperties;

    reviewCardLightClassName?: string;
    reviewCardLightStyle?: React.CSSProperties;

    reviewCardDarkClassName?: string;
    reviewCardDarkStyle?: React.CSSProperties;

    reviewBodyCardClassName?: string;
    reviewBodyCardStyle?: React.CSSProperties;

    reviewBodyTestimonialClassName?: string;
    reviewBodyTestimonialStyle?: React.CSSProperties;

    reviewTextClassName?: string;
    reviewTextStyle?: React.CSSProperties;

    reviewTextLightClassName?: string;
    reviewTextLightStyle?: React.CSSProperties;

    reviewTextDarkClassName?: string;
    reviewTextDarkStyle?: React.CSSProperties;

    reviewReadMoreClassName?: string;
    reviewReadMoreStyle?: React.CSSProperties;

    reviewReadMoreLightClassName?: string;
    reviewReadMoreLightStyle?: React.CSSProperties;

    reviewReadMoreDarkClassName?: string;
    reviewReadMoreDarkStyle?: React.CSSProperties;

    reviewFooterClassName?: string;
    reviewFooterStyle?: React.CSSProperties;

    reviewerClassName?: string;
    reviewerStyle?: React.CSSProperties;

    reviewerProfileClassName?: string;
    reviewerProfileStyle?: React.CSSProperties;

    reviewerProfileImageClassName?: string;
    reviewerProfileImageStyle?: React.CSSProperties;

    reviewerProfileFallbackClassName?: string;
    reviewerProfileFallbackStyle?: React.CSSProperties;

    reviewerNameClassName?: string;
    reviewerNameStyle?: React.CSSProperties;

    reviewerNameLightClassName?: string;
    reviewerNameLightStyle?: React.CSSProperties;

    reviewerNameDarkClassName?: string;
    reviewerNameDarkStyle?: React.CSSProperties;

    reviewerDateClassName?: string;
    reviewerDateStyle?: React.CSSProperties;

    reviewerDateLightClassName?: string;
    reviewerDateLightStyle?: React.CSSProperties;

    reviewerDateDarkClassName?: string;
    reviewerDateDarkStyle?: React.CSSProperties;
};

export type BadgeCSSProps = {
    badgeClassName?: string;
    badgeStyle?: React.CSSProperties;

    badgeContainerClassName?: string;
    badgeContainerStyle?: React.CSSProperties;

    badgeContainerLightClassName?: string;
    badgeContainerLightStyle?: React.CSSProperties;

    badgeContainerDarkClassName?: string;
    badgeContainerDarkStyle?: React.CSSProperties;

    badgeGoogleIconClassName?: string;
    badgeGoogleIconStyle?: React.CSSProperties;

    badgeInnerContainerClassName?: string;
    badgeInnerContainerStyle?: React.CSSProperties;

    badgeLabelClassName?: string;
    badgeLabelStyle?: React.CSSProperties;

    badgeLabelLightClassName?: string;
    badgeLabelLightStyle?: React.CSSProperties;

    badgeLabelDarkClassName?: string;
    badgeLabelDarkStyle?: React.CSSProperties;

    badgeRatingContainerClassName?: string;
    badgeRatingContainerStyle?: React.CSSProperties;

    badgeRatingClassName?: string;
    badgeRatingStyle?: React.CSSProperties;

    badgeRatingLightClassName?: string;
    badgeRatingLightStyle?: React.CSSProperties;

    badgeRatingDarkClassName?: string;
    badgeRatingDarkStyle?: React.CSSProperties;

    badgeStarsClassName?: string;
    badgeStarsStyle?: React.CSSProperties;

    badgeStarsContainerClassName?: string;
    badgeStarsContainerStyle?: React.CSSProperties;

    badgeStarsFilledClassName?: string;
    badgeStarsFilledStyle?: React.CSSProperties;

    badgeStarsEmptyClassName?: string;
    badgeStarsEmptyStyle?: React.CSSProperties;

    badgeLinkContainerClassName?: string;
    badgeLinkContainerStyle?: React.CSSProperties;

    badgeLinkClassName?: string;
    badgeLinkStyle?: React.CSSProperties;

    badgeLinkLightClassName?: string;
    badgeLinkLightStyle?: React.CSSProperties;

    badgeLinkDarkClassName?: string;
    badgeLinkDarkStyle?: React.CSSProperties;

    badgeLinkInlineClassName?: string;
    badgeLinkInlineStyle?: React.CSSProperties;
};

export type CarouselCSSProps = {
    carouselClassName?: string;
    carouselStyle?: React.CSSProperties;

    carouselBtnClassName?: string;
    carouselBtnStyle?: React.CSSProperties;

    carouselBtnLeftClassName?: string;
    carouselBtnLeftStyle?: React.CSSProperties;

    carouselBtnRightClassName?: string;
    carouselBtnRightStyle?: React.CSSProperties;

    carouselBtnLightClassName?: string;
    carouselBtnLightStyle?: React.CSSProperties;

    carouselBtnDarkClassName?: string;
    carouselBtnDarkStyle?: React.CSSProperties;

    carouselBtnIconClassName?: string;
    carouselBtnIconStyle?: React.CSSProperties;

    carouselCardClassName?: string;
    carouselCardStyle?: React.CSSProperties;
};

export type ErrorStateCSSProps = {
    errorClassName?: string;
    errorStyle?: React.CSSProperties;
};

export type LoadingStateCSSProps = {
    loaderClassName?: string;
    loaderStyle?: React.CSSProperties;

    loaderSpinnerClassName?: string;
    loaderSpinnerStyle?: React.CSSProperties;

    loaderLabelClassName?: string;
    loaderLabelStyle?: React.CSSProperties;
};